OpenSource cleanup and integration
  Merge our internal iOS branch to the main branch.

  Modified files:
    o sdk/objc/api/peerconnection/RTCHostname.mm
      In initWithHostname(), update.
    o sdk/objc/components/capturer/RTCCameraVideoCapturer.m
      Update comment.
    o sdk/objc/native/src/audio/audio_device_ios.mm
      Add missing cast to convert OSStatus into int in error message logging

+++ sdk/objc/native/src/audio/audio_device_ios.mm
@@ -691,7 +691,7 @@ void AudioDeviceIOS::HandleSampleRateChange() {
       [session notifyAudioUnitStartFailedWithError:result];
       RTCLogError(@"Failed to start audio unit with sample rate: %d, reason %d",
                   playout_parameters_.sample_rate(),
-                  result);
+                  (int) result);
       return;
     }
   }
@@ -891,7 +891,7 @@ void AudioDeviceIOS::UpdateAudioUnit(bool can_play_or_record) {
     OSStatus result = audio_unit_->Start();
     if (result != noErr) {
       [session notifyAudioUnitStartFailedWithError:result];
-      RTCLogError(@"Failed to start audio unit, reason %d", result);
+      RTCLogError(@"Failed to start audio unit, reason %d", (int)result);
       return;
     }
   }
